error pages

+++ Response.cpp
@@ -2,140 +2,6 @@
 #include "incs/Client.hpp"
 #include "Cgi.hpp"
 
-bool Response::send_informational_response() const {
-	return true;
-}
-
-bool Response::send_cgi_response(std::string body) const {
-
-	std::string response;
-
-	response = "HTTP/1.1 ";
-	response += toString(200);
-	response += " ";
-	response += this->get_code_msg(200);
-	response += "\r\n";
-	response += this->header_fields;
-	response += "\r\n";
-	response += body;
-	std::cerr << "response : "<< response <<"\n";
-	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
-		errMsgErrno("send failed");
-	return true;
-}
-bool Response::send_successful_response() const {
-
-	std::string response;
-
-	response = "HTTP/1.1 ";
-	response += toString(this->status_code);
-	response += " ";
-	response += this->get_code_msg(this->status_code);
-	response += "\r\n";
-	response += this->header_fields;
-	response += "\r\n";
-	response += this->body;
-	std::cerr << "RESponse :" << response <<"\n";
-	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
-		errMsgErrno("send failed");
-	return true;
-	// if (_code == 200)
-	// 	response = "HTTP/1.1 200 OK\n\nHello world\n";
-	// else if (_code == 201) {
-	// 	response = "HTTP/1.1 201 Created\nLocation: /\n\n";
-	// }
-}
-
-bool Response::send_redirection_message() const {
-	return true;
-}
-
-void Response::set_body(){
-	Location location = this->client.get_conf().get_location(this->client.get_request_target()).second;
-	std::string path = location.get_root() + this->client.get_request_target();
-	//std::cerr << "path" << path << ", " << this->client.get_request_target() << "\n";
-	std::string myline;
-
-	if (this->client.get_request_target().compare("/") == 0 ){
-		if (!location.get_index().empty())
-			path += location.get_index()[0];
-	}
-	std::cerr << "path.c_str():" << path.c_str() << "\n";
-	std::ifstream myfile (path.c_str());
-	if (myfile.is_open()){
-		while(myfile){
-			std::getline (myfile, myline);
-			this->body += myline;
-		}
-	}else {
-		this->status_code = 404;
-		std::cerr << "Error opening file\n";
-	}
-	//std::cerr << "body :" <<this->body << "\n";
-}
-
-bool Response::send_client_error_response() const {
-
-	//manque header
-	// NOTE : Search into Config if there is default
-	// error pages set.
-	std::string response;
-	std::string myline;
-
-	response = "HTTP/1.1 ";
-	response += toString(this->status_code);
-	response += " ";
-	response += this->get_code_msg(this->status_code);
-	response += "\n";
-	std::ifstream myfile;
-	//response += "Connection: keep-alive\nCache-Control: no-cache\nContent-Type: text/html;charset=UTF-8\nKeep-Alive: timeout=20\nVary: Accept-Encoding\nPragma: no-cache\nX-NWS-LOG-UUID: bd27210a-24e5-4740-8f6c-25dbafa9c395\nContent-Length: 180945\n\r\n";
-	if (this->status_code == 403)
-		std::ifstream myfile ("tests/www/errors/403.html");
-	else if (this->status_code == 405)
-		std::ifstream myfile ("tests/www/errors/405.html");
-	else if (this->status_code % 100 == 5)
-		std::ifstream myfile ("tests/www/errors/500.html");
-	else 
-		std::ifstream myfile ("tests/www/errors/404.html");
-	if (myfile.is_open()){
-		while(myfile){
-			std::getline (myfile, myline);
-			response += myline;
-		}
-	}
-	std::cerr << "ERROR response :" <<response << "\n";
-	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
-		errMsgErrno("send failed");
-	return true;
-}
-
-bool Response::send_server_error_response() const {
-
-	std::string response;
-	if (this->client.get_code() == 500)
-		response = "HTTP/1.1 500 Internal Server Error\n";
-	else if (this->client.get_code() == 501)
-		response = "HTTP/1.1 501 Not Implemented\n";
-	else if (this->client.get_code() == 505)
-		response = "HTTP/1.1 505 HTTP Version Not Supported\n";
-
-	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
-			errMsgErrno("send failed");
-	return true;
-}
-
-Response::Response(Client client): client(client){
-
-	std::cerr << "get_request_target :" << client.get_request_target()<< "\n";
-	std::cerr << "get_path2 :" << client.get_path2()<< "\n";
-	this->status_code_list = init_code_msg();
-	this->http_version = "HTTP/1.1";
-	this->status_code = 0;
-}
-Response::~Response(void){
-
-}
-
 void Response::check_setting_location(Location location, Config conf){
 	//check allow method
 	if (location.is_method_allowed(this->client.get_method()) == false){
@@ -164,7 +30,7 @@ void Response::check_setting_location(Location location, Config conf){
 	//if root exist check by access
 
 }
-void Response::set_header_fields(int cont_Leng, Location check_location){
+void Response::set_header_fields(int cont_Leng, Location check_location) {
 	std::map<std::string, std::string> headers;
 
 	headers["Content-Length"] = toString(cont_Leng);
@@ -237,6 +103,119 @@ bool Response::delete_file(){
 	}
 	return true;
 }
+
+bool Response::send_informational_response()  {
+	return true;
+}
+
+bool Response::send_cgi_response(std::string body) const  {
+
+	std::string response;
+
+	response = "HTTP/1.1 ";
+	response += toString(200);
+	response += " ";
+	response += this->get_code_msg(200);
+	response += "\r\n";
+	response += this->header_fields;
+	response += "\r\n";
+	response += body;
+	std::cerr << "Response :"<< response <<"\n";
+	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
+		errMsgErrno("send failed");
+	return true;
+}
+bool Response::send_successful_response()  {
+
+	std::string response;
+
+	response = "HTTP/1.1 ";
+	response += toString(this->status_code);
+	response += " ";
+	response += this->get_code_msg(this->status_code);
+	response += "\r\n";
+	response += this->header_fields;
+	response += "\r\n";
+	response += this->body;
+	std::cerr << "Response :" << response <<"\n";
+	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
+		errMsgErrno("send failed");
+	return true;
+	// if (_code == 200)
+	// 	response = "HTTP/1.1 200 OK\n\nHello world\n";
+	// else if (_code == 201) {
+	// 	response = "HTTP/1.1 201 Created\nLocation: /\n\n";
+	// }
+}
+
+bool Response::send_redirection_message()  {
+	return true;
+}
+
+void Response::set_body(){
+	Location location = this->client.get_conf().get_location(this->client.get_request_target()).second;
+	std::string path = location.get_root() + this->client.get_request_target();
+	//std::cerr << "path" << path << ", " << this->client.get_request_target() << "\n";
+	std::string myline;
+
+	if (this->client.get_request_target().compare("/") == 0 ){
+		if (!location.get_index().empty())
+			path += location.get_index()[0];
+	}
+	std::cerr << "path.c_str():" << path.c_str() << "\n";
+	std::ifstream myfile (path.c_str());
+	if (myfile.is_open()){
+		while(myfile){
+			std::getline (myfile, myline);
+			this->body += myline;
+		}
+	}else {
+		std::cerr << "Error opening file\n";
+
+	}
+	//std::cerr << "body :" <<this->body << "\n";
+}
+
+bool Response::send_error_response(Location location) {
+
+	std::string response;
+	std::string body;
+
+	response = "HTTP/1.1 ";
+	response += toString(this->status_code);
+	response += " ";
+	response += this->get_code_msg(this->status_code);
+	response += "\r\n";
+
+	body = "<!DOCTYPE html><html lang=\"en\"><head><title>";
+	body += toString(this->status_code) + " " + this->get_code_msg(this->status_code);
+	body += "</title></head><body><center><h1>";
+	body += toString(this->status_code) + " " + this->get_code_msg(this->status_code);
+	body +="</h1></center></body></html>";
+
+	set_header_fields(body.size(), location);
+	response += this->header_fields;
+	response += "\r\n";
+	response += body;
+
+	std::cerr << "ERROR response :" <<response << "\n";
+	if (send(this->client.get_fd(), response.c_str(), response.size(), 0) < 0)
+		errMsgErrno("send failed");
+	return true;
+}
+
+Response::Response(Client client): client(client){
+
+	std::cerr << "get_request_target :" << client.get_request_target()<< "\n";
+	std::cerr << "get_path2 :" << client.get_path2()<< "\n";
+	this->status_code_list = init_code_msg();
+	this->http_version = "HTTP/1.1";
+	this->status_code = 0;
+}
+Response::~Response(void){
+
+}
+
 bool Response::send_response(){
 	// NOTE : TO DEV
 	// Returns true if we need to close the connection
@@ -249,7 +228,7 @@ bool Response::send_response(){
 	this->check_setting_location(check_location, conf);
 	
 	//conf.debug();
-	if(check_location.get_cgi().first == true) {
+	if(check_location.get_cgi().first == true && this->status_code == 0) {
 		std::pair<bool, std::string> cgi_body;
 		std::string reponse;
 		//location is cgi;
@@ -262,21 +241,20 @@ bool Response::send_response(){
 			set_header_fields(cgi_body.second.size(), check_location);
 			return send_cgi_response(cgi_body.second);
 		}
-	}else if(this->client.get_method() == "GET"){
+	}else if(this->client.get_method() == "GET" && this->status_code == 0){
 			set_body();
 			set_header_fields(this->body.size(), check_location);
 			if (!this->status_code)
 				this->status_code = 200;
-	}else if (this->client.get_method() == "DELETE"){
+	}else if (this->client.get_method() == "DELETE" && this->status_code == 0){
 			if (delete_file() == false && this->status_code == 0)
 				this->status_code = 501;
 			else if (!this->status_code)
 				this->status_code = 202;
 	}
-	if (this->status_code >= 400 && this->status_code <= 499)
-		return send_client_error_response();
-	else if (this->status_code >= 500 && this->status_code <= 599)
-		return send_server_error_response();
+	//redicetion teste??
+	if (this->status_code >= 400 && this->status_code <= 599)
+		return send_error_response(check_location);
 	return send_successful_response();
 	// }else if method  get{
 	// 	if autoindex()
